+++ lib/public/components/qcFlags/isRunSubjectToQc.js
@@ -20,7 +20,7 @@ const QcDisabledReason = {
 };
 
 /**
- * Set of names of tags that prevent run to be subject to QC
+ * Set of names of tags which prevent run to be subject to QC
  */
 const RUN_TAGS_DISABLING_QC = new Set(['Not for physics']);
