+++ lib/public/components/qcFlags/qcFlagsPagesButtons.js
@@ -13,7 +13,7 @@
 
 import { frontLink } from '../common/navigation/frontLink.js';
 import { tooltip } from '../common/popover/tooltip.js';
-import { getRunNotSubjectToQcReason, isRunNotSubjectToQc } from './badRunInAsyncQcMessage.js';
+import { getRunNotSubjectToQcReason, isRunNotSubjectToQc } from './isRunSubjectToQc.js';
 import { h, iconPlus, iconWarning } from '/js/src/index.js';
 
 /**
