+++ lib/public/views/QcFlags/Create/qcFlagCreationComponent.js
@@ -23,7 +23,7 @@ import { formatRunStart } from '../../Runs/format/formatRunStart.js';
 import { formatRunEnd } from '../../Runs/format/formatRunEnd.js';
 import spinner from '../../../components/common/spinner.js';
 import errorAlert from '../../../components/common/errorAlert.js';
-import { getRunNotSubjectToQcReason, isRunNotSubjectToQc } from '../../../components/qcFlags/badRunInAsyncQcMessage.js';
+import { getRunNotSubjectToQcReason, isRunNotSubjectToQc } from '../../../components/qcFlags/isRunSubjectToQc.js';
 
 /**
  * Render the creation form of QC flag
