+++ lib/public/views/Runs/ActiveColumns/runDetectorsAsyncQcActiveColumns.js
@@ -17,7 +17,7 @@ import { qcFlagCreationPanelLink, qcFlagOverviewPanelLink } from '../../../compo
 import { VIRTUAL_DETECTOR_NAME } from '../../../domain/enums/detectorsNames.mjs';
 import { QC_SUMMARY_COLORS } from '../../../components/qcFlags/qcSummaryColors.js';
 import { tooltip } from '../../../components/common/popover/tooltip.js';
-import { getRunNotSubjectToQcReason, isRunNotSubjectToQc } from '../../../components/qcFlags/badRunInAsyncQcMessage.js';
+import { getRunNotSubjectToQcReason, isRunNotSubjectToQc } from '../../../components/qcFlags/isRunSubjectToQc.js';
 
 /**
  * Render QC summary for given run and detector
