+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewPage.js
@@ -26,7 +26,7 @@ import { inelasticInteractionRateActiveColumnsForProtonProton } from '../ActiveC
 import { filtersPanelPopover } from '../../../components/Filters/common/filtersPanelPopover.js';
 import runNumberFilter from '../../../components/Filters/RunsFilter/runNumber.js';
 import { qcSummaryLegendTooltip } from '../../../components/qcFlags/qcSummaryLegendTooltip.js';
-import { isRunNotSubjectToQc } from '../../../components/qcFlags/badRunInAsyncQcMessage.js';
+import { isRunNotSubjectToQc } from '../../../components/qcFlags/isRunSubjectToQc.js';
 
 const TABLEROW_HEIGHT = 59;
 // Estimate of the navbar and pagination elements height total; Needs to be updated in case of changes;
